+++ packages/lib/src/lib/map/floors/Floor.tsx
@@ -7,14 +7,18 @@ import {
   type LabelsMap,
   type LabelText,
   type OsmRenderMapProps,
-} from '../../types'
-import { type BoxBox } from '../box/prefixed'
-import type { Cb } from '../cb'
-import { floor_appearing_animation } from '../css'
-import { useLayout2 } from '../style/style-react'
-import { useFloors, type UseFloorsReturn } from '../viewer/floors/floors-react'
-import { useFloorRef } from '../viewer/floors/style'
-import { MAP_SVG_FLOORS } from './map-svg-react'
+} from '../../../types'
+import { type BoxBox } from '../../box/prefixed'
+import type { Cb } from '../../cb'
+import { floor_appearing_animation } from '../../css'
+import { useStyleRef } from '../../style/ref'
+import { useLayout2 } from '../../style/style-react'
+import {
+  useFloors,
+  type UseFloorsReturn,
+} from '../../viewer/floors/floors-react'
+import { useFloorRef } from '../../viewer/floors/style'
+import { MAP_SVG_FLOORS } from '../map-svg-react'
 
 export function RenderFloors(props: Readonly<OsmRenderMapProps>): ReactNode {
   return (
@@ -243,37 +247,64 @@ text, tspan {
 }
 */
 
-function RenderFloorHtmlLabels({ labels }: Props): ReactNode {
+function RenderFloorHtmlLabels({ idx, labels }: Props): ReactNode {
   return (
     <div className="labels">
-      {(labels ?? []).map((_text, idx) => {
+      {(labels ?? []).map((_text, lidx) => {
         return (
-          <div
-            key={idx}
-            className="label"
-            style={{
-              transform: `
+          <Fragment key={lidx}>
+            <RenderFloorHtmlLabel _text={_text} fidx={idx} idx={lidx} />
+          </Fragment>
+        )
+      })}
+      <style>{htmlLabelsStyle}</style>
+    </div>
+  )
+}
+
+const labelRefs = new Map()
+
+function RenderFloorHtmlLabel({
+  fidx,
+  _text,
+  idx,
+}: Readonly<{
+  fidx: number
+  _text: LabelText
+  idx: number
+}>): ReactNode {
+  const ref = useRef(null)
+  useStyleRef(labelRefs, ref, `f${fidx}-label-${idx}`)
+  return (
+    <div
+      style={{
+        transform: `
 translate(${_text.attrs?.['x'] || 0}px,${_text.attrs?.['y'] || 0}px)
+`,
+      }}
+    >
+      <div
+        ref={ref}
+        className="label"
+        style={{
+          transform: `
 rotate(${_text.attrs?.['rotate'] || 0}deg)
 scale(${_text.attrs?.['scale2'] || 0})
 scale(${_text.attrs?.['scale1'] || 0})
 translate(-50%, -50%)
 `,
-            }}
-          >
-            {'id' in _text.attrs && 'style' in _text.attrs && (
-              <style>{`#${_text.attrs['id']} { ${_text.attrs['style']}; }`}</style>
-            )}
-            {_text.children &&
-              _text.children.map((_tspan, idx2) => (
-                <p key={idx2} {...fromAttrs(_tspan.attrs)}>
-                  {_tspan.text ?? ''}
-                </p>
-              ))}
-          </div>
-        )
-      })}
-      <style>{htmlLabelsStyle}</style>
+        }}
+      >
+        {'id' in _text.attrs && 'style' in _text.attrs && (
+          <style>{`#${_text.attrs['id']} { ${_text.attrs['style']}; }`}</style>
+        )}
+        {_text.children &&
+          _text.children.map((_tspan, idx2) => (
+            <p key={idx2} {...fromAttrs(_tspan.attrs)}>
+              {_tspan.text ?? ''}
+            </p>
+          ))}
+      </div>
     </div>
   )
 }
